refactor: extract data hooks and components
